-------------------------------------------------------    nasciiboy 05-01, 05-0[3-6], 05-12, 05-1[5-9] ------------------------------------------------------- Sonic-Boom -------------------------------------------------------    Archive ! Restriction            ! Noise            : Take My Head -------------------------------------------------------

+++ 03-05/nasciiboy/main.go
@@ -84,7 +84,7 @@ const secs   = 3               // sections of color
 const sec    = colors / secs   // length of section
 
 func MagicColor( x, scale int ) color.RGBA {
-  i := int(x * colors / scale)
+  i := x * colors / scale
 
   h := float64(i % sec) / sec
   s := 1.0
